Flytt kode som beregner ordinær barnetrygd for barna ut i egen fil
